Version 0.2.2: - Added collection methods to relation collection - Small bugfixes - Fixed non working test module (Believer::Test::TestRunLifeCycle)

+++ lib/believer/test/test_run_life_cycle.rb
@@ -12,34 +12,27 @@ module Believer
         end
       end
 
-      def saved_models
-        @saved_models ||= []
-      end
-
-      def after_save(model)
-        saved_models << model
-      end
-
       # Detroys all CqlRecord::Base instances created
       class Destructor < Believer::Observer
         observe Believer::Base
 
         def cleanup
-          saved_models.each do |model|
-            begin
-              model.destroy
-            rescue Exception => e
-              puts "Could not destroy model #{model}: #{e}\n#{e.backtrace.join("\n")}"
+          unless @saved_models.nil? || @saved_models.empty?
+            puts "Cleaning up #{@saved_models.size} objects"
+            @saved_models.each do |model|
+              begin
+                model.destroy
+              rescue Exception => e
+                puts "Could not destroy model #{model}: #{e}\n#{e.backtrace.join("\n")}"
+              end
             end
+            @saved_models = nil
           end
         end
 
-        def saved_models
-          @saved_models ||= []
-        end
-
         def after_save(model)
-          saved_models << model
+          @saved_models ||= []
+          @saved_models << model
         end
 
       end
